triangles change

+++ Triangle.java
@@ -6,9 +6,9 @@
 public class Triangle {
 	public static void main(String[] args)
 	{
-		double a = Double.parseDouble(args[0]);
-		double b = Double.parseDouble(args[1]);
-		double c = Double.parseDouble(args[2]);
+		int a = Integer.parseInt(args[0]);
+		int b = Integer.parseInt(args[1]);
+		int c = Integer.parseInt(args[2]);
 
 		boolean isTriangle = (a + b > c) && (a + c > b) && (b + c > a);
 		System.out.println(a + ", " + b + ", " + c + ": " + isTriangle);
